feat: authPage 리펙토링 (Agentic-AI-Busan#72)
